feat: auth 에 따라 마이페이지, 로그인 모달 보여주기
